requirements.txt
# Core dependencies
pyspark==3.5.0
delta-spark==3.0.0
pyyaml==6.0.1
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
pydantic==2.4.2

# Databricks
databricks-sdk==0.18.0

# Configuration and utilities
python-dotenv==1.0.0
pytz==2023.3.post1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
